refactor(network): extract NetworkView sections into components
